testcase moved to service package
